% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{add_features}
\alias{add_features}
\title{Add features like stage and tumor type to an initialized population}
\usage{
add_features(popbysim, probs, names = NULL)
}
\arguments{
\item{popbysim}{A matrix of dimensions (population size)x(number of sims)}

\item{probs}{A vector of probabilities summing to 1. Each probability
represents the probability of being in each feature group
stage and tumor-type subgroup}

\item{stagetumor}{Optional vector of names corresponding to the 
stage-tumor subgroups in probs. Defaults to 1:length(probs)}
}
\value{
Matrix of dimensions (pop size)x(nsim) indicating stage-tumor 
status according to the 'names' parameter
Defaults to 1:length(probs)
}
\description{
Add features like stage and tumor type to an initialized population
}
\examples{
Say there are four stage-ER receptor subgroups that have equal probability,
in a population of size 20 and 2 simulations
add_features(matrix(1, nrow=20, ncol=2), probs=c(0.25, 0.25, 0.25, 0.25))
Could ID the groups by name
add_features(matrix(1, nrow=20, ncol=2), probs=c(0.25, 0.25, 0.25, 0.25),
             names=c('EarlyER+', 'EarlyER-', 'AdvancedER+', 'AdvancedER-'))

}

