% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{allmortrates}
\alias{allmortrates}
\title{Format life table data from IHME or other source}
\usage{
allmortrates(lifefile, source = "ihme")
}
\arguments{
\item{lifefile}{Data frame with lifetable data}

\item{source}{Data source, only 'ihme' supported right now}
}
\value{
Formatted all-cause mortality rates by age
}
\description{
Format life table data from IHME or other source
}
