% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{bcirates}
\alias{bcirates}
\title{Format female breast cancer incidence rates from SEER or IARC/GLOBOCAN
Specifies age by the midpoint of the age category}
\usage{
bcirates(incfile, source = "seer")
}
\arguments{
\item{incfile}{Path to file}

\item{source}{One of 'seer' or 'globocan'}
}
\value{
Data frame
}
\description{
Format female breast cancer incidence rates from SEER or IARC/GLOBOCAN
Specifies age by the midpoint of the age category
}
\examples{
IARC/Globocan data
bcirates(system.file('extdata', 
                    'incidence/CI5-Xd/18000299_uga.csv', package='bcimodel'),
        source='globocan')

IARC/Globocan data
bcirates(system.file('extdata', 
                    'incidence/SEER/bc_1975-1979_incidence.csv', 
                     package='bcimodel'),
        source='globocan')

}

