% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cantrance.R
\name{calc_ac_lifespan}
\alias{calc_ac_lifespan}
\title{Generate an age at other-cause death from a lifetable}
\usage{
calc_ac_lifespan(ageentry, male, lifetable, n_sim = 100, time = FALSE,
  haz = 1, max100 = TRUE)
}
\arguments{
\item{ageentry}{An integer age}

\item{male}{Sex indicator: 1 for male, 0 for female}

\item{lifetable}{Data frame with columns Survival, BirthCohort, Age, and Male}

\item{n_sim}{Number of random draws to return}

\item{time}{Set to TRUE to return time between ageentry and death, instead of age at death}

\item{haz}{Hazard to apply to the life table survival}

\item{max100}{If TRUE, interpolate out to 100 as the max survival age}
}
\value{
A vector of length n_sim containing ages at other-cause death (or times to other-cause death, if time=TRUE)
}
\description{
Given a year of birth and current age, draws a random age at other-cause death from a lifetable
}
\examples{
library(bcimodel)
data(allmortratesf)
surv <- interpolate_cumsurv(allmortratesf, 
                          ratevar='Rate.Per.100K',
                          country='Uganda')
# Code compatibility tweaks for life table
surv <- transform(surv, Age=age, Survival=cumsurv, Male=0)
# Generate 5 death times for a female of age 30
calc <- ac <- lifespan(30, 0, surv, 5)

}
