% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{calcarr}
\alias{calcarr}
\title{Calculate absolute risk reduction}
\usage{
calcarr(cummortlist, controlindex, reverse = FALSE)
}
\arguments{
\item{cummortlist}{List of cummort() outputs}

\item{controlindex}{Number indicating which cummortlist element is the control scenario}

\item{reverse}{Set to TRUE to return -1*ARR}
}
\value{
Matrix of ARRs for each scenario/cummortlist element
}
\description{
For a list of scenarios, calculate absolute risk reductions (ARRs) compared to the control scenario
}
\examples{

}

