% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{calcmrr}
\alias{calcmrr}
\title{Calculate mortality rate ratios}
\usage{
calcmrr(cummortlist, controlindex, reverse = FALSE, perc = FALSE)
}
\arguments{
\item{cummortlist}{List of cummort() outputs}

\item{controlindex}{Number indicating which cummortlist element is the control scenario}

\item{reverse}{Set to TRUE to return 1-MRR}

\item{perc}{Set to TRUE to report a percent, instead of a ratio}
}
\value{
Matrix of MRRs for each scenario/cummortlist element
}
\description{
For a list of scenarios, calculate mortality rate ratios (MRRs) compared to the control scenario
}
\examples{

}

