% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{compile_naturalhist}
\alias{compile_naturalhist}
\title{Creates a 'naturalhist' object that describes key natural history parameters
for incident cases}
\usage{
compile_naturalhist(prop_adv, mortrates, subgroup_probs)
}
\arguments{
\item{prop_adv}{Advanced Proportion of cancers presenting as Advanced stage}

\item{mortrates}{Named vector of cancer mortality rates by stage,
e.g. c(Early=.05, Advanced=0.21)}

\item{subgroup_probs}{Named vector of subgroup probabilities that occur within stage groups, e.g. c(`ER+`=0.5, `ER-`=0.5)}
}
\value{
Data frame of class 'naturalhist' with columns stage, subgroup,
mortrate and prop
}
\description{
Takes in stats on cancer stage, subgroup (e.g. tumor type) and survival
and compiles into a data frame of class 'naturalhist'. Note that this function
could be modified to allow for more complex situations. Right now it presumes
that mortality rates vary by stage only, and that stage and subgroup are
uncorrelated.
}
\examples{
compile_naturalhist(prop_adv=0.85, mortrates=c(Early=0.05, Advanced=0.21), 
                   subgroup_probs=c(`ER+`=0.5, `ER-`=0.5))

}

