% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earlydetect.R
\name{create_stageshift_map}
\alias{create_stageshift_map}
\title{Creates a "map" of how the stage shift will be applied}
\usage{
create_stageshift_map(x)
}
\arguments{
\item{x}{Data frame of class naturalhist (see compile_naturalhist)}
}
\value{
Matrix that maps which rows of the naturalhist data frame 
correspond to Early vs Advanced stages
}
\description{
Follows the principle that when an advanced cases is shifted to early stage
via early detection, the subgroup (e.g., tumor type) remains constant.
}
\examples{
nh <- compile_naturalhist(prop_adv=0.85, mortrates=c(Early=0.05, Advanced=0.21), 
                   subgroup_probs=c(`ER+`=0.5, `ER-`=0.5))
create_stageshift_map(nh)

}

