% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{cummort}
\alias{cummort}
\title{Tally cumulative incidence of cancer mortality}
\usage{
cummort(futimes, times, cancerdeath)
}
\arguments{
\item{futimes}{Vector of follow-up times to assess}

\item{times}{Matrix of times to cancer mortality}

\item{cancerdeath}{Logical matrix indicating death from cancer}
}
\value{
Matrix of mortality counts for sims (rows) and follow-up times (columns)
}
\description{
Count cancer mortality within a follow-up period
}
\examples{

}

