% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{cumsurv_to_exprate}
\alias{cumsurv_to_exprate}
\title{Convert survivals to rates}
\usage{
cumsurv_to_exprate(cum.surv, year = 10, haz = 1)
}
\arguments{
\item{cum.surv}{Proportion surviving}

\item{year}{Time corresponding to cum.surv, i.e. time since time=0, cum.surv=1}

\item{haz}{Hazard ratio, of the cum.surv reflects an original curve raised to a hazard}
}
\description{
Convert survivals to rates
}

