% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{cumyears}
\alias{cumyears}
\title{Tally years lived within a follow-up period}
\usage{
cumyears(futimes, times)
}
\arguments{
\item{futimes}{Vector of follow-up times to assess}

\item{times}{Matrix of times to cancer mortality}
}
\value{
Matrix of cumulative years lived within the follow-up period for sims (rows) and follow-up times (columns)
}
\description{
Count years alive within a follow-up period, summed over all individuals per sim
}
\examples{
# Assess 5- and 10-year follow-ups
fu <- c(5, 10)
times <- matrix(rexp(n=25, rate=1/5), nrow=5, ncol=5)
cumyears(fu, times)

}

