% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{format_age}
\alias{format_age}
\title{Convert age-group population sizes to population proportions}
\usage{
format_age(age_data, gsize = 5, minAge = 20, maxAge = 85,
  format = FALSE)
}
\arguments{
\item{age_data}{Population age structure data frame OR file}

\item{gsize}{Number of years covered by each age group}

\item{minAge}{Optional lower age limit for the population}

\item{maxAge}{Optional upper age limit for the population}

\item{format}{Does the input need to be formatted? If so, 
age_data can be a .csv file}
}
\description{
Takes population counts by age group and returns the population proportions
for single-year age groups
}
\note{
See bcimodel/data-raw/agestructure.R
}
