% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{format_bounds_list}
\alias{format_bounds_list}
\title{Return bounds for each element of the list
Options are: lower, upper; (lower, upper); mean (lower, upper)}
\usage{
format_bounds_list(thislist, digits = NULL, paren = FALSE,
  includemean = FALSE, compileall = FALSE)
}
\description{
Return bounds for each element of the list
Options are: lower, upper; (lower, upper); mean (lower, upper)
}

