% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{interpolate_cumsurv}
\alias{interpolate_cumsurv}
\title{Compute cumulative event-free survival curve for single-year ages}
\usage{
interpolate_cumsurv(binnedrates, ratevar, maxage = NULL, country = NULL)
}
\arguments{
\item{binnedrates}{Data frame with Age for age groups}

\item{ratevar}{Name of rate variable}

\item{maxage}{Maximum age desired. If data are incomplete, the 
rate for the oldest age will be held constant until maxage}

\item{country}{Country to select from the data frame}
}
\value{
Data frame of cumulative survival probability by single-year ages
}
\description{
Takes age-binned event rates per 100,000 and uses interpolation to estimate
cumulative event-free survival for single years
}
\examples{
data(incratesf)
data(allmortratesf)

inc <- interpolate_cumsurv(incratesf, 
                         ratevar='Female.Rate.Per.100K',
                         country='Uganda')
mort <- interpolate_cumsurv(allmortratesf, 
                         ratevar='Rate.Per.100K',
                         country='Uganda')

}

