% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelized.R
\name{parsim_treatment_by_subgroup}
\alias{parsim_treatment_by_subgroup}
\title{Simulate treatment by subgroups - parallelized
Given the proportions of treatments by subgroups, simulate treatment}
\usage{
parsim_treatment_by_subgroup(treat_chars, stage_subgroup_rows, thisprop,
  pop_size, nsim, ncores = 4)
}
\arguments{
\item{treat_chars}{Data frame indicating stage-subgroup #s in the SSno 
variable and stage-subgroup-treatment #s in txSSno, and 
prop_* columns where * is the trial name.}

\item{stage_subgroup_rows}{Matrix indicating, for each person-sim, 
which stage-subgroup
they are in (using the SSid from treat_chars)}

\item{thisprop}{Character indicating the prop_* column name from which
to take treatment proportions}

\item{pop_size}{Population size}

\item{nsim}{Number of sims}

\item{ncores}{Number of cores}
}
\description{
Simulate treatment by subgroups - parallelized
Given the proportions of treatments by subgroups, simulate treatment
}
