% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpolicies.R
\name{parsimpolicies}
\alias{parsimpolicies}
\title{Same as simpolicies but calling parallelized sub-functions}
\usage{
parsimpolicies(scenarios, naturalhist, treatinfo, agesource = "Standard",
  minage = 0, maxage = 100, incsource = "Uganda", mortsource = "Uganda",
  popsize = 1e+05, sims = 5, futimes = c(5, 10),
  returnstats = c("mean"), denom = 1e+05, ncores = 4)
}
\value{
Table if returnstats='mean', or list of tables of length(returnstats)
}
\description{
Same as simpolicies but calling parallelized sub-functions
}

