% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelized.R
\name{partreatments_by_policy}
\alias{partreatments_by_policy}
\title{Use sim_treatment_by_subgroup to simulate treatments, parallelized}
\usage{
partreatments_by_policy(policies, treat_chars, stagegroups, map, pop_size, nsim,
  ncores = 4, method = "inner")
}
\arguments{
\item{policies}{A "scenarios" data frame containing an 'id' for the policies
and a 'pairnum' column indicating either NA or the paired policy, for}

\item{method}{Either 'outer' or 'inner'. If outer, will parallelize the outer
lapply. If inner, will parallelize repeated calls to sim_multinom}
}
\description{
Simulate treatments according to specified policy rules
}

