% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{plot_rates}
\alias{plot_rates}
\title{Plot the data in the incidence database}
\usage{
plot_rates(x, y, psize = NULL, group = NULL)
}
\arguments{
\item{x}{Probably age group indicator (must be numeric)}

\item{y}{Probably age-specific rates (per 100,000)}

\item{psize}{Variable to scale point size}

\item{group}{Variable indicating groups to plot using different colors}
}
\value{
ggplot
}
\description{
Plot the data in the incidence database
}
\examples{
# Plot incidence
data(incratesf)
plot_rates(incratesf$Age, incratesf$Female.Rate.Per.100K,
        psize=incratesf$Cases, group=paste(incratesf$Country,
                                           incratesf$Year))

# Plot mortality
data(allmortratesf)
df <- subset(allmortratesf, 
                  Country=='United States' | Country=='Uganda')
plot_rates(df$Age, df$Rate.Per.100K, psize=NULL, group=df$Country)

}

