% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{plot_results}
\alias{plot_results}
\title{Plot results
Select on measures and panel them. Group on f-u year?}
\usage{
plot_results(rlong, measures = NULL, type = "bar", rotate_xlab = TRUE)
}
\description{
Plot results
Select on measures and panel them. Group on f-u year?
}

