% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inc_and_lifet.R
\name{rate_to_cumsurv}
\alias{rate_to_cumsurv}
\title{Given event rates, compute cumulative survival}
\usage{
rate_to_cumsurv(x)
}
\arguments{
\item{x}{Vector of event rates (interpreted as discrete rates, not instantaneous hazards)}
}
\value{
Vector of cumulative survivals
}
\description{
Given event rates, compute cumulative survival
}

