% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{return_value_from_id}
\alias{return_value_from_id}
\title{Look up the value that corresponds to a numeric ID}
\usage{
return_value_from_id(ids, df, value)
}
\arguments{
\item{ids}{Matrix of ids}

\item{df}{Data frame containing the values of interest, sorted by ID (i.e., row numbers should correspond to the ID numbers)}

\item{value}{Column name of value to return from the df}
}
\value{
Matrix of values corresponding to the input ID matrix
}
\description{
Using a matrix of IDs, look up a value in a dataframe
}
\examples{
# Simple age distribution
ages <- data.frame(age=c(20,30,40), prop=c(0.2, 0.5, 0.3))
# Simulate a population of size 10 twice, using these proportions and the row IDs
sims <- sim_multinom(nsims=100, 2, ages$prop, names=1:nrow(ages))
# Get the corresponding ages
sim.ages <- return_value_from_id(sims, df=ages, value='age')
# Proportions match input props
round(prop.table(table(sim.ages)), 2)

}
