% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{returnstat4matrix}
\alias{returnstat4matrix}
\title{Compute mean and/or quantiles
Return mean, lower, or upper quantile. Default quantile type is 7, which gives a continuous result}
\usage{
returnstat4matrix(matrix, stat, quanttype = 7)
}
\arguments{
\item{matrix}{Matrix of values to summarize}

\item{stat}{Statistic to return: 'mean', 'lower' or 'upper'}

\item{quanttype}{Quantile type - see R help page for quantile()}
}
\description{
Compute mean and/or quantiles
Return mean, lower, or upper quantile. Default quantile type is 7, which gives a continuous result
}

