% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inc_and_lifet.R
\name{running_product}
\alias{running_product}
\title{Return running product of a vector, for computing cumulative survival}
\usage{
running_product(x)
}
\arguments{
\item{x}{Vector of conditional survival probabilities}
}
\value{
Vector of running products
}
\description{
Return running product of a vector, for computing cumulative survival
}

