% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment.R
\name{shifttreatment_indicator}
\alias{shifttreatment_indicator}
\title{Create indicator of whether treatment should be re-simulated}
\usage{
shifttreatment_indicator(x, type, shifts, basecase, map)
}
\arguments{
\item{x}{Numeric ID}

\item{type}{Vector of either NA or numeric ID referring to scenario number that is}

\item{shifts}{List of shift matrices}

\item{basecase}{Base case matrix of stage-subgroup indicators}

\item{map}{Map indicating stage-subgroup pairs}
}
\description{
Determines TRUE or FALSE indicator based on scenario settings.
}
\examples{

}

