% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{sim_multinom}
\alias{sim_multinom}
\title{Simulate a category assignment using a multinomial distribution}
\usage{
sim_multinom(nsims, nreps, probs, names)
}
\arguments{
\item{nsims}{Number of size=1 sims from the multinomial}

\item{nreps}{Number of times to replicate the nsims. Put the smaller number here, the bigger one as nsims}

\item{probs}{Vector of probabilities for each of the K categories}

\item{names}{Vector of names associated with each of the categories. Must be of same length as probs. Can be character or numeric. Use numeric row IDs if you're referring to a group defined by multiple variables whose groupings are defined in a separate data frame}
}
\value{
Matrix of multinomial draws
}
\description{
Simulate a category assignment using a multinomial distribution
}
\examples{
sim_multinom(10, 5, c(0.1, 0.3, 0.6), names=c('a', 'b', 'c'))
sim_multinom(5, 10, c(0.1, 0.3, 0.6), names=c('a', 'b', 'c'))

}

