% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpolicies.R
\name{simpolicies}
\alias{simpolicies}
\title{Run the model for a series of policies and compare outcomes to the base case
Wrapper function for the whole model}
\usage{
simpolicies(scenarios, naturalhist, treatinfo, agesource = "Standard",
  minage = 0, maxage = 100, incsource = "Uganda", mortsource = "Uganda",
  popsize = 1e+05, sims = 5, futimes = c(5, 10),
  returnstats = c("mean"), denom = 1e+05)
}
\arguments{
\item{scenarios}{Data frame of scenarios to simulate (see define_scenarios). First scenario should be the base case.}

\item{naturalhist}{Data frame with natural history parameters (see compile_naturalhist)}

\item{agesource}{Country to use for age structure (see data(agestructure) )}

\item{minage}{Lower age limit for population at sim start}

\item{maxage}{Upper age limit for population at sim start}

\item{incsource}{Country to use for incidence rates (see data(incratesf) )}

\item{mortsource}{Country to use for life table (see data(allmortratesf) )}

\item{popsize}{Size of population to simulate}

\item{sims}{Number of simulations}

\item{futimes}{Follow-up times at which to tally outcomes, in years}

\item{returnstats}{Defaults to simulation means, but you can also ask for c("mean", "lower", "upper"). Lower and upper report the 2.5\% and 97.5\% quantiles respectively.}

\item{Denominator/population}{size by which to report outcomes; default is 100,000}
}
\value{
Data frame if returnstats="mean", or list of data frames of length(returnstats)
}
\description{
Run the model for a series of policies and compare outcomes to the base case
Wrapper function for the whole model
}
\examples{
# Use example input data loaded in package, object name "ex1"
data(ex1) 
ex1

# Element "pol" shows the 3 scenarios being modeled
"nh" contains the natural history information
"map" contains numeric indicators for the stage-subgroups
"tx" shows treatment hazard ratios and, for each scenario, the treatment distribution. See vignette for more details

# Run a model using ex1 inputs and the defaults
uganda_stdpop <- simpolicies(ex1$pol, ex1$nh, ex1$tx)

}

