% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earlydetect.R
\name{stageshift_indicator}
\alias{stageshift_indicator}
\title{Sim indicator of stageshift from binomial}
\usage{
stageshift_indicator(x, pop_size, nsim)
}
\arguments{
\item{x}{Hazard ratio for early detection (i.e., 1-proportion stage shifted)}

\item{pop_size}{Population size}

\item{nsim}{Number of sims}
}
\value{
Matrix of 1s and 0s
}
\description{
Sim indicator of stageshift from binomial
}
\examples{
#

}

