% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment.R
\name{treatments_by_policy}
\alias{treatments_by_policy}
\title{Use sim_treatment_by_subgroup to simulate treatments}
\usage{
treatments_by_policy(policies, treat_chars, stagegroups, map, pop_size, nsim)
}
\arguments{
\item{policies}{A "scenarios" data frame containing an 'id' for the policies
and a 'pairnum' column indicating either NA or the paired policy, for}
}
\description{
Simulate treatments according to specified policy rules
}

