% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment.R
\name{update_treat_stageshift}
\alias{update_treat_stageshift}
\title{For stage-shifted cases in an early detection scenario, update treatment}
\usage{
update_treat_stageshift(policies, shifts, treats)
}
\description{
Takes treatments from a paired scenario with no early detection and 
updates treatments only for those advanced-stage cases who were
stage-shifted in the early detection scenario
}

