% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{rexp_matrix}
\alias{rexp_matrix}
\title{Sim from exponential given rates in matrix form}
\usage{
rexp_matrix(ratematrix)
}
\description{
This is simple code but it's a separate function to confirm that
it performs as expected
}

