% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpolicies.R
\name{simpolicies}
\alias{simpolicies}
\title{Run the model for a series of policies and compare outcomes to the base case}
\usage{
simpolicies(scenarios, naturalhist, treatinfo, agesource = "Standard",
  minage = 0, maxage = 100, incsource = "Uganda", mortsource = "Uganda",
  popsize = 1e+05, sims = 5, futimes = c(5, 10),
  returnstats = c("mean"), denom = 1e+05)
}
\arguments{
\item{scenarios}{Data frame of scenarios to simulate (see define_scenarios). First scenario should be the base case.}

\item{naturalhist}{Data frame with natural history parameters (see compile_naturalhist)}

\item{agesource}{Country to use for age structure (see data(agestructure) )}

\item{minage}{Lower age limit for population at sim start}

\item{maxage}{Upper age limit for population at sim start}

\item{incsource}{Country to use for incidence rates (see data(incratesf) )}

\item{mortsource}{Country to use for life table (see data(allmortratesf) )}

\item{popsize}{Size of population to simulate}

\item{sims}{Number of simulations}

\item{futimes}{Follow-up times at which to tally outcomes}

\item{returnstats}{Defaults to means across sims, but you can also ask for c('mean', 'lower', upper')}

\item{Denominator}{by which to report outcomes
male=0, i.e. all female sex.}
}
\value{
Table if returnstats='mean', or list of tables of length(returnstats)
}
\description{
Run the model for a series of policies and compare outcomes to the base case
}
\examples{
# Use example input data
uganda_stdpop <- simpolicies(ex1$pol, ex1$nh, ex1$tx)

}

