% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{timetocancerdeath_by_policy}
\alias{timetocancerdeath_by_policy}
\title{Sim time to cancer death based on policy guidelines}
\usage{
timetocancerdeath_by_policy(policies, rates, pop_size, nsim)
}
\arguments{
\item{policies}{A "scenarios" data frame containing an 'id' for the policies
and a 'pairnum' column indicating either NA or the paired policy, for}
}
\description{
Sim non-stage-shifted times, to be updated for shifted cases using the 
update_time_stageshift function
}

