% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{update_time_stageshift}
\alias{update_time_stageshift}
\title{For stage-shifted cases in an early detection scenario, update treatment}
\usage{
update_time_stageshift(policies, shifts, rates, times)
}
\description{
Takes treatments from a paired scenario with no early detection and
updates treatments only for those advanced-stage cases who were
stage-shifted in the early detection scenario
}

